@@ -1,16 +1,13 @@
 {
   "name": "slacky-slack",
   "version": "1.0.0",
-  "engines": {
-    "node": "6.10.2"
-  },
+
   "description": "Prebuilt api for slack app clone.",
-  "main": "dist",
+  "type": "module",
+  "main": "./src/index.js",
+  
   "scripts": {
-    "dev": "NODE_ENV=development nodemon -w src --exec \"babel-node src --presets es2015,stage-0\"",
-    "build": "babel src -s -D -d dist --presets es2015,stage-0",
-    "prestart": "npm run -s build",
-    "start": "node ./dist/index.js",
+    "start": "node --experimental-modules src/index.js",
     "lint": "eslint src",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -27,26 +24,26 @@
       "no-unsed-vars": 1
     }
   },
-  "author": "Jacob Luetzow",
+  "author": "updated by akimov.infotech",
   "license": "ISC",
   "dependencies": {
-    "body-parser": "1.17.1",
-    "express": "4.15.2",
-    "express-jwt": "5.3.0",
-    "jsonwebtoken": "7.4.0",
-    "mongoose": "4.9.7",
-    "passport": "0.3.2",
+    "body-parser": "1.19.0",
+    "express": "4.16.4",
+    "express-jwt": "5.3.1",
+    "jsonwebtoken": "8.5.1",
+    "mongoose": "5.5.6",
+    "passport": "0.4.0",
     "passport-local": "1.0.0",
-    "passport-local-mongoose": "4.0.0",
-    "babel-cli": "6.24.1",
-    "babel-core": "6.24.1",
-    "babel-eslint": "7.2.3",
+    "passport-local-mongoose": "5.0.1",
+    "babel-cli": "6.26.0",
+    "babel-core": "6.26.3",
+    "babel-eslint": "10.0.1",
     "babel-preset-es2015": "6.24.1",
     "babel-preset-stage-0": "6.24.1",
-    "socket.io": "2.0.3"
+    "socket.io": "2.2.0"
   },
   "devDependencies": {
-    "eslint": "3.19.0",
-    "nodemon": "1.11.0"
+    "eslint": "5.16.0",
+    "nodemon": "1.19.0"
   }
 }
