@@ -1,18 +1,18 @@
 import jwt from 'jsonwebtoken';
 import expressJwt from 'express-jwt';
 
-const TOKENTIME = 60*60*24*90;
+const TOKENTIME = 60 * 60 * 24 * 90;
 const SECRET = "W3 CHAT 4 FUN";
 
 let authenticate = expressJwt({ secret: SECRET });
 
 let generateAccessToken = (req, res, next) => {
   req.token = req.token || {};
-  req.token = jwt.sign ({
+  req.token = jwt.sign({
     id: req.user.id,
   }, SECRET, {
-    expiresIn: TOKENTIME // 90 days
-  });
+      expiresIn: TOKENTIME // 90 days
+    });
   next();
 }
 
@@ -23,8 +23,4 @@ let respond = (req, res) => {
   });
 }
 
-module.exports = {
-  authenticate,
-  generateAccessToken,
-  respond
-}
+export { authenticate, generateAccessToken, respond }
